requirements-dev.txt
black==23.3.0
mypy==1.4.1
pre-commit==3.3.3
watchdog
pytest
streamlit
pyyaml
pandas
transformers
torch
streamlit
requests
python-dotenv